Move ascend-to-package-dir
