+++ packages/schema/src/scripts/util/ascend-to-package-dir.ts
